config/requirements.txt
asgiref==3.5.2
Django==3.2
django-colorfield==0.7.2
django-countries==7.4.2
django-phonenumber-field==7.0.0
django-resized==1.0.2
phonenumbers==8.13.0
Pillow==9.3.0
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
